detect-tls-cert-serial: move unittests to tests/
